🔄 refactor(ci): centralize CI logic in core and streamline imports
Moved CI utilities to core for better modularity and reuse.
