+++ packages/cli/src/nodehost.ts
@@ -63,7 +63,7 @@ import { resolveLanguageModel } from "../../core/src/lm"
 import { CancellationOptions } from "../../core/src/cancellation"
 import { defaultModelConfigurations } from "../../core/src/llms"
 import { createPythonRuntime } from "../../core/src/pyodide"
-import { ci } from "./ci"
+import { ci } from "../../core/src/ci"
 import { arrayify } from "../../core/src/cleaners"
 import { McpClientManager } from "../../core/src/mcpclient"
 import { ResourceManager } from "../../core/src/mcpresource"
