+++ packages/core/src/ci.ts
@@ -1,0 +1,5 @@
+import _ci from "ci-info"
+
+export const ci = _ci
+
+export const isCI = _ci.isCI
